Refactor DNSCheckWidget: Extract business logic to custom hook
Changes:
- Created useDNSCheck custom hook in src/hooks/useDNSCheck.ts
- Hook encapsulates all business logic:
  * State management (status, random strings, PC check state)
  * SSE connection handling
  * Fetch request logic
  * Timeout management
  * Cleanup on unmount
  * Browser check and PC check implementations
- Refactored DNSCheckWidget to use the hook
- Component now focused only on presentation logic
- Improved dialog close handling (resets state if closed during waiting)

Benefits:
- Separation of concerns (presentation vs business logic)
- Easier to test (hook can be tested in isolation)
- More maintainable and readable code
- Component reduced from ~300 to ~180 lines
- Hook logic can be reused if needed

The widget functionality remains exactly the same, just with better code organization.
